requirements.txt
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.10.0
colorama==0.4.4
decorator==5.1.1
executing==0.8.3
ipython==8.1.1
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.22.2
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.28
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
traitlets==5.1.1
wcwidth==0.2.5
